feat: add filter for author and categories
