+++ generate/templates/brand.ejs
@@ -1,5 +1,6 @@
 import { Author, getCredit } from "../credits";
 import { IBrand } from "../type";
+import { BrandCategory } from "../categories";
 
 export const <%= className %>: IBrand = {
     name: '<%= name %>',
