+++ src/data/ProgrammingVTuberLogos/IntelliJIDEA.ts
@@ -1,9 +1,11 @@
+import { BrandCategory } from "../categories";
 import { Author, getCredit } from "../credits";
 import { IBrand } from "../type";
 
 export const IntelliJIDEA: IBrand = {
   name: "IntelliJ IDEA",
   aliases: ["JetBrains", "jb"],
+  categories: [BrandCategory.IDE],
   logos: [
     {
       url: "https://raw.githubusercontent.com/Aikoyori/ProgrammingVTuberLogos/main/IntelliJIDEA/IntellijLogo.png",
