+++ src/data/ProgrammingVTuberLogos/VSCode.ts
@@ -1,9 +1,12 @@
+import { BrandCategory } from "../categories";
 import { Author, getCredit } from "../credits";
 import { IBrand } from "../type";
 
 export const VSCode: IBrand = {
   name: "Visual Studio Code",
   aliases: ["vsc", "vscode", "vs code"],
+  categories: [BrandCategory.IDE],
+
   logos: [
     {
       url: "https://raw.githubusercontent.com/Aikoyori/ProgrammingVTuberLogos/main/VSCode/VSCode.png",
