+++ src/data/ServiceLogos/Angular.ts
@@ -1,8 +1,10 @@
+import { BrandCategory } from "../categories";
 import { Author, getCredit } from "../credits";
 import { IBrand } from "../type";
 
 export const Angular: IBrand = {
   name: "Angular",
+  categories: [BrandCategory.FRAMEWORK],
   logos: [
     {
       url: "https://raw.githubusercontent.com/SAWARATSUKI/ServiceLogos/main/Angular/Angular.png",
