+++ src/data/ServiceLogos/C.ts
@@ -1,8 +1,11 @@
+import { BrandCategory } from "../categories";
 import { Author, getCredit } from "../credits";
 import { IBrand } from "../type";
 
 export const C: IBrand = {
   name: "C",
+  categories: [BrandCategory.LANGUAGE],
+
   logos: [
     {
       url: "https://raw.githubusercontent.com/SAWARATSUKI/ServiceLogos/main/C/C.png",
