+++ src/data/ServiceLogos/FORTRAN.ts
@@ -1,8 +1,11 @@
+import { BrandCategory } from "../categories";
 import { Author, getCredit } from "../credits";
 import { IBrand } from "../type";
 
 export const FORTRAN: IBrand = {
   name: "FORTRAN",
+  categories: [BrandCategory.LANGUAGE],
+
   logos: [
     {
       url: "https://raw.githubusercontent.com/SAWARATSUKI/ServiceLogos/main/FORTRAN/FORTRAN.png",
