+++ src/data/ServiceLogos/Flutter.ts
@@ -1,8 +1,11 @@
+import { BrandCategory } from "../categories";
 import { Author, getCredit } from "../credits";
 import { IBrand } from "../type";
 
 export const Flutter: IBrand = {
   name: "Flutter",
+  categories: [BrandCategory.FRAMEWORK],
+
   logos: [
     {
       url: "https://raw.githubusercontent.com/SAWARATSUKI/ServiceLogos/main/Flutter/FlutterTransparent.png",
