+++ src/data/ServiceLogos/Html.ts
@@ -1,8 +1,11 @@
+import { BrandCategory } from "../categories";
 import { Author, getCredit } from "../credits";
 import { IBrand } from "../type";
 
 export const Html: IBrand = {
   name: "Html",
+  categories: [BrandCategory.LANGUAGE],
+
   logos: [
     {
       url: "https://raw.githubusercontent.com/SAWARATSUKI/ServiceLogos/main/Html/HTML.png",
