+++ src/data/ServiceLogos/Rstudio.ts
@@ -1,8 +1,9 @@
 import { Author, getCredit } from "../credits";
 import { IBrand } from "../type";
-
+import { BrandCategory } from "../categories";
 export const Rstudio: IBrand = {
   name: "Rstudio",
+  categories: [BrandCategory.IDE],
   logos: [
     {
       url: "https://raw.githubusercontent.com/SAWARATSUKI/ServiceLogos/main/Rstudio/RStudio.png",
