+++ src/data/ServiceLogos/Ruby.ts
@@ -1,8 +1,11 @@
+import { BrandCategory } from "../categories";
 import { Author, getCredit } from "../credits";
 import { IBrand } from "../type";
 
 export const Ruby: IBrand = {
   name: "Ruby",
+  categories: [BrandCategory.LANGUAGE],
+
   logos: [
     {
       url: "https://raw.githubusercontent.com/SAWARATSUKI/ServiceLogos/main/Ruby/Ruby.png",
